test(vllm): isolate CLI tests from backend-free runners
