+++ components/src/dynamo/vllm/tests/test_vllm_backend_args.py
@@ -14,7 +14,6 @@ from types import SimpleNamespace
 
 import pytest
 
-from dynamo.vllm.args import parse_args
 from dynamo.vllm.backend_args import (
     DisaggregationMode,
     DynamoVllmArgGroup,
@@ -465,41 +464,6 @@ class TestClassifyWorkerExclusivity:
         config.benchmark_mode = "agg"
         config.headless = True
         config._validate_classify_worker_exclusivity()
-
-
-@pytest.mark.usefixtures("vllm_cpu_platform_when_no_accelerator")
-class TestParseArgsLoraExclusivity:
-    """The --enable-lora exclusivity rules must fire on the real CLI path.
-
-    The tests above call the validators directly with engine_args already set,
-    so they pass even when parse_args never supplies engine_args before
-    validate() runs. These drive the whole command line instead, which is the
-    only place that ordering is observable.
-    """
-
-    @staticmethod
-    def _parse(extra_argv):
-        return parse_args(["--model", "Qwen/Qwen3-0.6B", *extra_argv])
-
-    def test_realtime_with_enable_lora_is_rejected(self):
-        with pytest.raises(ValueError, match="enable-lora"):
-            self._parse(["--realtime", "--enable-lora"])
-
-    def test_classify_worker_with_enable_lora_is_rejected(self):
-        """Kept separate from the --realtime case: the classify rule may be
-        removed once LoRA is supported on pooling-family workers, and the
-        --realtime rule is independent of that."""
-        with pytest.raises(ValueError, match="enable-lora"):
-            self._parse(["--classify-worker", "--enable-lora"])
-
-    def test_enable_lora_alone_is_accepted(self):
-        config = self._parse(["--enable-lora"])
-        assert config.engine_args.enable_lora is True
-
-    def test_realtime_alone_is_accepted(self):
-        config = self._parse(["--realtime"])
-        assert config.realtime is True
-        assert not config.engine_args.enable_lora
 
 
 class TestValidateCustomEncoder:
